Preliminary fix for x86.

+++ libgbaic/src/input_file.cpp
@@ -358,7 +358,9 @@ void input_file::convert_to_binary(elfio& reader)
                 {
                     // Move to start of segment in output. Fill up with padding bytes.
                     const auto nbytes = current->get_virtual_address() - output_address;
-                    m_data.insert(m_data.end(), nbytes, 0); // TODO: this is broken when building for x86. Fix it.
+                    // TODO: cast needed for x86. For the sake of paranoia, don't blindly cast. Check if nbytes fits into size_t first.
+                    //       If it doesn't, throw.
+                    m_data.insert(m_data.end(), static_cast<size_t>(nbytes), 0);
                     output_address += nbytes;
                 }
                 else
